+++ src/silk/placeholder.py
@@ -1,7 +1,8 @@
-from fp_ops.placeholder import Placeholder, _ # noqa
+from fp_ops.placeholder import Placeholder # noqa
 # just to make the placeholder available in the namespace
 
 
 
 
 
+_ = Placeholder()